# Backend_API_NAES/Backend_API_NAES/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Backend_API_NAES, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Tarjeta": {
      "get": {
        "tags": [
          "Tarjeta"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Tarjeta"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/TarjetaCredito"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Tarjeta/{id}": {
      "put": {
        "tags": [
          "Tarjeta"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/TarjetaCredito"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Tarjeta"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "TarjetaCredito": {
      "required": [
        "cvv",
        "fechaExpiracion",
        "numeroTarjeta",
        "titular"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "titular": {
          "minLength": 1,
          "type": "string"
        },
        "numeroTarjeta": {
          "minLength": 1,
          "type": "string"
        },
        "fechaExpiracion": {
          "minLength": 1,
          "type": "string"
        },
        "cvv": {
          "minLength": 1,
          "type": "string"
        }
      }
    }
  }
}